Remove some "as usize" which are now no longer needed

+++ src/main.rs
@@ -358,7 +358,7 @@ fn setup_new_game(
 ) {
     // Update GameUi with current map
     let mut game_ui = ui::components::GameUi::new();
-    game_ui.set_map(cgd.maps.map[cgd.current_map as usize].clone());
+    game_ui.set_map(cgd.maps.map[cgd.current_map].clone());
     uicomps.comps.insert(
         GAME_UI_NAME.to_string(),
         UIComponentData {
@@ -435,7 +435,7 @@ fn update_map(cgd: Res<CurrentGameData>, mut uicomps: ResMut<UIComponents>, quer
 
     // Update Game UI
     let mut game_ui = ui::components::GameUi::new();
-    game_ui.set_map(cgd.maps.map[cgd.current_map as usize].clone());
+    game_ui.set_map(cgd.maps.map[cgd.current_map].clone());
     game_ui.set_entities(result);
     uicomps.comps.insert(
         GAME_UI_NAME.to_string(),
